[SPARK-44266][SQL] Move Util.truncatedString to sql/api
